♻️(dateManipulator): move into lib
